Finish Command pattern, add deprecation notices
